Fixed sonar

+++ src/main/java/g3301_3400/s3304_find_the_k_th_character_in_string_game_i/Solution.java
@@ -12,7 +12,7 @@ public class Solution {
             s = sb.toString();
             char[] cq = s.toCharArray();
             for (char c1 : cq) {
-                int ascii = (int) c1;
+                int ascii = c1;
                 c1 = (char) (ascii + 1);
                 sb.append(c1);
             }
